Simplify kern feature generation a bit
Adding subtables to feature syntax is unnecessary because feaLib will
do this itself.

Checking for conflicts in class-to-class rules is unnecessary because
conflicts from other rules are considered exceptions and are allowed.
In fact, we can probably do without checking for conflicts anywhere
(if we're checking already in glyphs2ufo) except that we enumerate
class-to-glyph rules. It would be best to do without this enumeration
but that seems to cause problems with compilation (this is something
to look into).