[kernWriter] Don't use feaTools
This is a pretty trivial piece of code to include, and it's one less
dependency.

The new matching scheme shares one limitation with feaTools: it
doesn't understand the concept of glyph ranges. This might be a
problem for removing kerning conflicts, though I'd like to remove
conflict detection from here as well (leaving it up to the user).