[feaLib] use feaLib's addOpenTypeFeaturesFromString function
feaLib's API is in constant change. See https://github.com/behdad/fonttools/pull/556

In the latest fontTools HEAD, the function `addOpenTypeFeatures` now takes either a
path or a file-like object as the input featurefile.

A new `addOpenTypeFeaturesFromString` was added to parse features from a string --
which is what ufo2ft wants to use. The 'filename' argument is optional in this case,
as it's only used for error messages and relative include paths (a bit unusual
for a feature files contained inside a UFO folder).