[mark] Support mkmk and mark2liga by default
This expands the mark feature writer to generate mark-to-ligature
rules, and creates an mkmk feature by default.

Mark class definitions are moved outside of any feature, since they
are likely to be shared by the mark and mkmk features. This means that
a lot of code had to be moved around.

The "combAccentOnly" and "checkAliases" arguments are of dubious
value, but Roboto depends on them for now. Ideally that can be changed
in the Roboto sources themselves, and then this code can be cleaned up
a bit.