outlineOTF: do not allow negative advance widths when generating hmtx table
The hmtx specs clearly say `advanceWidth` field in `_longHorMetric` struct must be
an `USHORT`.

FreeType also interprets the field as an unsigned short, leading to extremely wide
rendering.

Acrobat Reader returns a "A number is out of range" error if a negative advance is
found in a PDF, and then all the widths will show up scrambled.

The UFO spec is silent about it (just says 'integer or float'), however Glyphs.app
prevents this issue by not allowing to set negative advance widths.

Hence we should do the same in ufo2ft.