fontInfoData: use Python2 round() for postscriptWeightNameFallback
Again this is to make output consistent across 2 and 3.

Here I choose the Python 2 variant because the result are large numbers,
(multiples of 100), and there's a high probability that input values
happen to be in the middle. Using the banker's rounding would produce
unexpected results.

E.g. we want 550 to be rounded to 600, and 650 to 700, i.e. away from 0
like in Python 2. In Python 3 both will be rounded 'half-even-up' to 600:

>>> from fontTools.misc.py23 import round2, round3
>>> round2(550, -2)
>>> 600.0
>>> round2(650, -2)
>>> 700.0
>>> round3(550, -2)
>>> 600
>>> round3(650, -2)
>>> 600