[outlineCompiler] fix bbox calculation; support roundTolerance and floats in CFF
This should fix issues #24 and #106.

The OutlineOTFCompiler now takes a 'roundTolerance' float argument that is passed on to the fontTools T2CharStringPen and affects how float coordinates are rounded to integers.
The default value (0.5) means round all coordinates to integers, 0 means no rounding, whereas numbers in between result in some rounding and not others.

The font and glyphs bounding box calculation has been refactored to account for the new CFF roundTolerance parameter, and also to fix issue #26, whereby the use of math.floor() instead of round() when computing the bounds of TrueType glyphs could lead to the hmtx's left sidebearings be out-of-sync with the glyf's xMin values.

The hhea and vhea are now computed directly from the values in hmtx and vmtx tables, rather than independently from them. The control point bounding box is used instead of the defcon glyph's {left,right,top,bottom}Margin attributes, as the latter are based on the actual glyph bounds, whereas in the {h,v}mtx we use the control point bounds.