[fontInfoData] use preferred family and subfamily to compose the full name
instead of the styleMap* names.

The opentype spec allows both:

> The full font name is generally a combination of name IDs 1 and 2,
or of name IDs 16 and 17, or a similar human-readable variant.

I tend to prefer the "preferred" ones, as the styleMap* names can
sometimes be shortened to fit the legacy max-32 char limitations,
while the full name should have more room to accommodate the *full*
names...

There is still the open question as to whether to keep or leave the
"Regular" in the full name.
I'm currently swinging towards keeping it...