[markFeatureWriter] ensure pos rule order is determistic
when one does `for glyph in font`, the `__iter__` method yields
glyph objects in random order because font.keys() is an unsorted
collection in defcon.

Thus, when we are building the mark feature and collecting anchors
we have to iterate over glyphs in a specific order, i.e. the
font.glyphOrder if it's present, else alphabetic order.