read private 'useProductionNames' key from UFO lib
This is the same as what fontmake currently does with the
Glyphs.app's specific "Don't use Production Names", only that
it is made a positive boolean value for clarity's sake.

The default behaviour of ufo2ft doesn't change. It still does
the post/CFF renaming by default. However when this new lib
key "com.github.googlei18n.ufo2ft.useProductionNames" is present
and set to False, the renaming is disabled.