compiler_test: rm Bug108.ufo/kerning.plist
current kernFeatureWriter does not skip generating kern feature even if it's already
present. It appends to it by default. We will test this behaviour separately.
Here we only test that the included file is parsed, and that the
relative path is resolved relative to the UFO path, not the features.fea
file.