[outlineCompiler] don't set CID-only FontName operator in CFF TopDict
The 'FontName' operator is only valid in CID-keyed CFF fonts

Since we emit a name-keyed CFF table, this field is redundant at best.

Even worse, on latest macOS 10.13 the presence of this operator in the
CFF breaks the QuickLook preview (space-bar from Finder), and the font
appears all blank; maybe they use it to detect whether CFF is name- or
CID-keyed, and then fail to render because in fact it isn't CID...

Besides, MakeOTF silently strips this when it is passed an OTF produced
by fontmake.

So we can safely remove it.

Note that the postscriptName from the fontinfo.plist is still encoded
in the Name INDEX structure (ttx dumps this as the 'name' attribute of
the CFFFont element).
It is just the dict operator that should not be there.