kernFeatureWriter: make DFLT kern lookup when splitting LTR/RTL
This 'kern_dflt' lookup includes all the pairs for the neutral-direction
glyphs (Common or Inherited script), which as such do not belong to
neither the LTR scripts' group, nor the RTL one.
The lookup is included in all defined languagesystems.
If we are not splitting up LTR/RTL lookups, then we do not need to
create a separate kern_dflt lookup.
Note that this lookup is only registered for the kern feature, and
does not affect the dist feature (if any).

https://github.com/googlei18n/ufo2ft/issues/198#issuecomment-391699814