Add _GlyphSet layer-dict adapter object
basically a dict with a 'lib' attribute.

GlyphSets are expected to be dict objects in the ufo2ft API,
whereas defcon (and ufoLib2) Layer objects only expose part of the
MutableMapping interface.
Besides, in cu2qu fonts_to_quadratic the 'fonts' are expected to have
a 'lib' attribute where the 'curve_type' value is stored.

Also, we remove unused copyGlyphSet func (it wasn't part of public API)