Drop unused dependency from requirements
The `requirements.txt` file appears to be out of sync with the dependencies listed in `setup.py`. A searched the project and I don't see any reference to this being used, but I could easily be wrong about that. If this fix is wrong then the correct thing to do with be to add the requirement to `setup.py` so they are at least in sync.