filters/featureWriters: use '...' as placeholder for loading from UFO lib
Currently the `filters` and `featureWriters` parameters override the ones defined in the UFO lib.plist, and it's impossible to extend instead of replace these, e.g. inserting some additional filter or writer either before or after.

This PR allows to pass a special value `...` (the actual ``ellipsis`` singleton, not the str literal '...') in both these lists, which acts as a placeholder to signal one wants to extend the list of predefined filters/writers.

E.g.:

filters=[..., DecomposeTransformedComponentsFilter(pre=True)]

The ellipsis is replaced by the custom filters defined in the UFO lib (if any), and the additional one is appended to the list.
Only one ellipsis is allowed in the list.

Fontmake will also be updated to allow passing --filter='...' or --feature-writer='...' CLI options to use this feature.

Part of fixing https://github.com/googlefonts/fontmake/issues/872