Add feaIncludeDir option to allow override the include search path
This is intended to fix https://github.com/googlefonts/glyphsLib/issues/797, by allowing fontmake to override the default include directory used for parsing a UFO features.fea; useful when the UFOs where generated from a .glyphs source whose features contain some `include` statements and they are saved in a different directory (or not saved to disk at all).

fontmake will need to be modified to call ufo2ft with this new option when starting to build from a .glyphs input.