Improve mark writer for character used in multiple scripts
If a character is used in multiple scripts, and on of these scripts uses
e.g. USE shaper, the mark writer will write its anchors in abvm/blwm
features exclusively. This breaks the mark attachment if the shaper of
the indented script does not enable abvm/blwm features.

This is a partial fix, by looking into the language systems declared in
the feature file and filtering out abvm scripts not in them.

Part of https://github.com/googlefonts/ufo2ft/issues/579