InstructionCompiler: move autoUseMyMetrics method to this, where it's used
By default we only try to automatically set the USE_MY_METRICS flags if there are no UFO components that contain the public.truetype.useMyMetrics lib key, so it is safe to have the option True by default.

Also, don't type the ufo and otf parameters as Optional, because they are not really optional, all uses assume that they are not none (besides some tests which I will fix).