Test that util function raises when multiple glyphs claim a code point
This is the documented behaviour of util.makeUnicodeToGlyphNameMapping()
but a typo means that the exception is not raised. This commit adds a
failing test to check for the documented behaviour, to confirm a future
fix and avoid regressions.

This commit also creates a file to contain tests for ufo2ft's util
functions to give this test a happy home.