[outlineCompiler] add empty glyphs for missing components for non-default TTF masters
For TrueType only, when the glyphSet for a non default master contains composite glyphs that point to missing component base glyphs, we create and add empty glyphs so that the composite glyph is not dropped from the master TTF as invalid (i.e. pointing to nowhere); varLib will ignore those empty glyphs when building gvar variations, so the additional glyphs will not add extra masters. The glyph metrics (HVAR, VVAR) similarly will not have additional masters for the empty glyphs (we use a sentinel value understood by varLib 'skip me').
The composite glyph, on the other hand, will still be considered when building variations.

One may use this technique to adjust the placement of components in a composite glyph, and/or its advance width/height, but only for some intermediate locations without requiring to define intermediate masters for the component glyphs themselves.