raise error (instead of warning) for missing components while decomposing
At some point we decided to skip missing components when decomposing composites and just issue a logging warning. But that was not a good idea, as it hides other font bugs (or even code bugs), and it is no longer needed because we now support sparse composite glyphs. We may reintroduce the skip behavior as an option if need be.