[kernFeatureWriter] ignore zero-valued class kern pairs when building variable kern
We already prune class-class kerning pairs when the value is 0 when building individual per-master GPOS tables (to be merged with varLib). It makes sense to also do that when we build a single variable GPOS table.
This is safe to do because we always place the class-class kerning as the last (after glyph-glyph kerns), so these don't override anything else that may follow and 0 is the default anyway, so they are essentially no-op.