markFeatureWriter: Ignore contextual anchors
We don’t support contextual anchors in markFeatureWriter, but glyphsLib
subclasses it and adds support for them. However, since we are unaware
of contextual anchors, we end up with positioning statements with
duplicated mark positions like this:

    pos base beh-ar.init
        <anchor 91 -316> mark @MC_top
        <anchor 91 -226> mark @MC_top;

(one is the regular anchor, and the other is the contextual one). Which
makes no sense (feaLib shouldn’t probably allow the same mark class to
be used multiple times in the same statement).

This PR adds isContextual to NamedAnchor and ignores any such anchors
when writing pose statements. It is already set by glyphsLib’s
MarkFeatureWriter, but otherwise should not affect ufo2ft’s as it always
sets it to False.