markFeatureWriter: Ignore contextual anchors in non-contextual lookups
Otherwise we end up with positioning statements that have duplicated
mark classes like this:

pos base a
   <anchor 200 200> mark @MC_top
   <anchor 100 200> mark @MC_top;

(one is the regular anchor, and the other is the contextual one). Which
makes no sense (feaLib shouldn’t probably allow the same mark class to
be used multiple times in the same statement).