Only include a format 4 cmap subtable when there are BMP codepoints
This will produce slightly smaller fonts when there are no characters
represented inside of the Unicode BMP (U+0000 to U+FFFF), and matches
fontations' current behaviour:

https://github.com/googlefonts/fontations/blob/9cc956edd1b4b36a1653c6faf341c9ed7dac753a/write-fonts/src/tables/cmap.rs#L192-L210

The OpenType specification states that a format 4 subtable is not
mandatory, but that it may be included for backwards compatibility for
legacy environments that do not support format 12. As in our case there
are no overlapping codepoints, this suggests we can drop the format 4
subtable entirely (at least, unless there has been software developed
since that unconditionally assumes its existence).

(a small aside after re-reading: in fontations and ufo2ft we are not
following the recommendation to include all BMP codepoints in the format
12 table too [!]; this does not seem to have caused problems though)

https://learn.microsoft.com/en-us/typography/opentype/spec/recom#cmap-table

In the case where this would result in 0 subtables, we still cannot drop
`cmap` entirely, as it is a required table.

https://learn.microsoft.com/en-us/typography/opentype/spec/otff#required-tables