[markFeatureWriter] Consolidate how we consider a mark contextual
A mark can be contextual if it starts with an asterisk and has
"GPOS_Context" in libData, otherwise it is just an ignorable anchor
(since it does not starts with a letter).

We were checking for LibData in some places, but in most places we only
check of isContextual. This change sets isContextual only of libData is
not None and it has "GPOS_Context" key.