kernFeatureWriter: Fix handling of neutral glyphs
We were returning the empty set when we did not have a known script for
a glyph, but the classify function explicitly expects `None` in this
case.

This was causing us to not correctly assign certain accent glyphs to the
appropriate script, and to instead generate an unnecessary extra lookup
for these glyphs.