preProcessor: decompose components whose transform overflows F2Dot14
Even though the TTGlyphPen in fonttools already does this for us, it is better to do it upfront so that filters that are performed later on in the preprocessing pipeline (e.g. removeOverlaps, cubicToQuadratic, etc.) operate on the decomposed outlines.

This also matches what fontc currently does, and as a bonus give us at least one +1 in the fontc_crater diff (i.e. Ruthie.glyphs).