[outlineCompiler] clamp xAvgCharWidth to avoid overflows when saving 'sparse' master TTFs
If a non-default, 'sparse' master does not contain a '.notdef', or if it
contains a composite glyph where some the component glyphs referenced
are missing from that master, we synthesize the missing glyphs and set their
advance width and height to a sentinel value 0xFFFF which varLib understands
as 'this glyph does not participate in {H,V}VAR glyph metrics variations'.
After fonttools computes their OS/2.xAvgCharWidth, the latter may end
up oveflowing the signed 16-bit integer limits.
Usually this is not a problem, because when compiling a VF with
fontmake -o variable, the TTFont masters are not compiled and saved to disk
but the objects are passed on to fontTools.varLib as is. But if one
calls fontmake -o ttf-interpolatable, all the master TTFs are compiled and
saved to disk and then a struct.error exception will be raised if the
xAvgCharWidth overflows.
The simplest thing to do is to clamp this computed (unused, useless) OS/2
table field to the range allowed by the spec and move on.