pyproject.toml: library dependencies should be un-pinned
When we replaced setup.py with pyproject.toml in https://github.com/googlefonts/ufo2ft/pull/951, the dependencies were inadvertently pinned to exact version, which is not what we want for a library like ufo2ft. The v3.6.7 release I made with this has been 'yanked' from PyPI and I will follow up with a new release that fixes this bug.

Sorry I did not catch this earlier.